+++ contracts/userInterfaces/tee/ITeeIdKeyIdPair.sol
@@ -4,5 +4,5 @@ pragma solidity >=0.7.6 <0.9;
 /// Tee id + key id pair structure
 struct TeeIdKeyIdPair {
     address teeId;
-    uint256 keyId;
+    uint64 keyId;
 }
